[REFACTOR] #85 Home 전체 구성 수정 전, 메뉴 테이블 존재
